Fix loss of accuracy of incoming parameter

+++ dmserver/src/display_manager_stub.cpp
@@ -251,7 +251,7 @@ int32_t DisplayManagerStub::OnRemoteRequest(uint32_t code, MessageParcel &data, 
         }
         case DisplayManagerMessage::TRANS_ID_SET_VIRTUAL_PIXEL_RATIO: {
             ScreenId screenId = static_cast<ScreenId>(data.ReadUint64());
-            float virtualPixelRatio = data.ReadUint32();
+            float virtualPixelRatio = data.ReadFloat();
             bool res = SetVirtualPixelRatio(screenId, virtualPixelRatio);
             reply.WriteBool(res);
             break;
